Big update: 3.4
- New: Redis support!
- Ability to convert old data to the new system
- Additional settings in settings.ini to allow using legacy server data
system
- Help message fix
- Added some common bots to ignore (to prevent circular commands)
- Bugfixes and adaptions to redis
- Various speed improvements

+++ plugins/reminder.py
@@ -3,9 +3,10 @@ import time
 import os
 import asyncio
 import logging
+import importlib
 from pickle import load, dumps
-from discord import Message, Client, DiscordException
-from data.utils import resolve_time, convert_to_seconds, is_valid_command, is_empty, StandardEmoji
+from discord import Message, Client, DiscordException, Object, User
+from data.utils import resolve_time, convert_to_seconds, is_valid_command, is_empty, StandardEmoji, decode, gen_id
 from data.stats import MESSAGE, WRONG_ARG
 
 log = logging.getLogger(__name__)
@@ -14,6 +15,11 @@ log.setLevel(logging.INFO)
 # CONSTANTS
 
 DEFAULT_REMINDER_LIMIT = 2
+TICK_DURATION = 15
+EXCEEDED_REMINDER_LIMIT = "{} You have exceeded the maximum amount of reminders (**{}** at once).".format(StandardEmoji.WARNING, DEFAULT_REMINDER_LIMIT)
+
+REMINDER_PERSONAL = "personal"
+REMINDER_CHANNEL = "channel"
 
 remind_help = "**Remind help**\n`_remind me in [sometime]: [message]` - reminds you in your DM\n" \
               "`_remind here in [sometime]: [message]` - reminds everyone in current channel"
@@ -33,14 +39,17 @@ valid_commands = commands.keys()
 # Functions
 
 
-class ReminderHandler():
-    def __init__(self, client, loop=asyncio.get_event_loop()):
+class LegacyReminderHandler:
+    def __init__(self, client, _, loop=asyncio.get_event_loop()):
         self.client = client
         self.loop = loop
 
         self.reminders = {}
 
         self.wait = False
+
+    def get_reminder_amount(self):
+        return len(self.reminders)
 
     def lock(self):
         self.wait = True
@@ -62,12 +71,12 @@ class ReminderHandler():
         if self.reminders.get(user.id):
             self.reminders[user.id] = []
 
-    def remove_reminder(self, user, reminder):
-        if self.reminders.get(user.id):
+    def remove_reminder(self, user_id, reminder):
+        if self.reminders.get(user_id):
 
-            for c, rm in enumerate(self.reminders[user.id]):
+            for c, rm in enumerate(self.reminders[user_id]):
                 if reminder == rm.get("raw"):
-                    self.reminders[user.id].pop(c)
+                    self.reminders[user_id].pop(c)
                     return True
 
         return False
@@ -160,6 +169,156 @@ class ReminderHandler():
             last_time = await self.tick(last_time)
 
 
+class RedisReminderHandler:
+    """
+    Datatype: Hash
+
+    reminder:<SERVER_ID> =>
+
+                    Key: Randomly generated if
+                    Value: Json-encoded
+                    --------------------------
+                    Values:
+                            full_time: Reminder length
+                            content: Formatted content to send
+                            receiver: User/Channel id
+                            time_created: unix epoch time
+                            time_target: time_created + full_time
+                            author: author id
+                            raw: raw content
+
+    """
+    def __init__(self, client, handler, loop=asyncio.get_event_loop()):
+        self.redis = handler.get_plugin_data_manager(namespace="reminder")
+
+        self.loop = loop
+        self.client = client
+        self.wait = False
+
+        try:
+            self.json = importlib.import_module("ujson")
+        except ImportError:
+            self.json = importlib.import_module("json")
+
+    def get_reminder_amount(self):
+        return len(self.get_all_reminders())
+
+    def is_active(self):
+        return bool(self.get_all_reminders())
+
+    def get_reminders(self, user_id):
+        if self.redis.exists(user_id):
+            return {int(idd): self.json.loads(a) for idd, a in self.redis.hgetall(user_id).items()}
+        else:
+            return []
+
+    def get_all_reminders(self):
+        return [self.get_reminders(decode(a).strip("reminder:")) for a in self.redis.scan_iter("*")]
+
+    def remove_all_reminders(self, user):
+        if self.redis.exists(user.id):
+            self.redis.delete(user.id)
+
+    def remove_reminder(self, user_id, reminder_id):
+        if self.redis.exists(user_id):
+            return self.redis.hdel(user_id, reminder_id)
+
+        return False
+
+    def check_reminders(self, user):
+        """
+        True == ok
+        False = not ok
+        :param user: User or Member
+        :return: success
+        """
+        if self.redis.exists(user.id):
+            return bool(len(self.get_reminders(user.id)) <= DEFAULT_REMINDER_LIMIT)
+
+        else:
+            return True
+
+    def set_reminder(self, channel, author, content, tim, reminder_type=REMINDER_PERSONAL):
+        """
+        Sets a reminder
+        :param channel: Where to send this
+        :param author: Who is the author
+        :param content: String : message
+        :param tim: time (int)
+        :param reminder_type: type of reminder (personal or channel)
+        :return: success
+        """
+        t = time.time()
+
+        if not self.check_reminders(author):
+            return -1
+
+        tim = convert_to_seconds(tim)
+        if not (5 <= tim < 172800):  # Allowed reminder duration: 5 sec to 2 days
+            return False
+
+        raw = str(content)
+
+        if reminder_type == REMINDER_PERSONAL:
+            template = "{} You asked me to remind you:\n```{}```"
+        else:
+            template = "{} Timer is up!\n```{}```"
+
+        content = template.format(StandardEmoji.ALARM, content)
+
+        # Add the reminder to the list
+        rm_id = gen_id(length=12)
+        tree = {"full_time": tim, "content": content, "receiver": channel.id, "time_created": int(t),
+                "time_target": int(tim + t), "author": author.id, "raw": raw, "type": reminder_type}
+        field = self.json.dumps(tree)
+
+        print("New reminder: {} to {}".format(raw, channel.id))
+
+        return self.redis.hset(author.id, rm_id, field)
+
+    @staticmethod
+    async def tick(last_time):
+        """
+        Very simple implementation of a self-correcting tick system
+        :return: None
+        """
+        current_time = time.time()
+        delta = TICK_DURATION - (current_time - last_time)
+
+        await asyncio.sleep(delta)
+
+        return time.time()
+
+    async def dispatch(self, rem):
+        try:
+            log.debug("Dispatching")
+
+            if rem.get("type") == REMINDER_PERSONAL:
+                receiver = User(id=rem.get("receiver"))
+            else:
+                receiver = Object(id=rem.get("receiver"))
+
+            await self.client.send_message(receiver, rem.get("content"))
+        except DiscordException as e:
+            log.warning(e)
+
+    async def start_monitoring(self):
+        last_time = time.time()
+
+        while True:
+            # Iterate through users and their reminders
+            for user in self.get_all_reminders():
+                for rm_id, reminder in user.items():
+                    # If enough time has passed, send the reminder
+                    if reminder.get("time_target") <= last_time:
+
+                        await self.dispatch(reminder)
+                        self.remove_reminder(reminder.get("author"), rm_id)
+
+            # And tick.
+            last_time = await self.tick(last_time)
+
+
 class Reminder:
     def __init__(self, **kwargs):
         self.client = kwargs.get("client")
@@ -167,29 +326,30 @@ class Reminder:
         self.handler = kwargs.get("handler")
         self.nano = kwargs.get("nano")
         self.stats = kwargs.get("stats")
+        self.legacy = kwargs.get("legacy")
 
-        self.reminder = ReminderHandler(self.client, self.loop)
+        if self.legacy:
+            self.reminder = LegacyReminderHandler(self.client, self.loop)
 
-        # Uses the cache if it exists
+            # Uses the cache if it exists
 
-        if os.path.isfile("cache/reminders.temp"):
+            if os.path.isfile("cache/reminders.temp"):
+                # Removes the file if it is empty
+                if is_empty("cache/reminders.temp"):
+                    os.remove("cache/reminders.temp")
+                else:
+                    log.info("Using reminders.cache")
 
-            # Removes the file if it is empty
-            if is_empty("cache/reminders.temp"):
-                os.remove("cache/reminders.temp")
+                    with open("cache/reminders.temp", "rb") as vote_cache:
+                        rem = load(vote_cache)
 
-            # or uses it
-            else:
-                log.info("Using reminders.cache")
+                    # 3.1.5 : disabled
+                    # os.remove("cache/reminders.temp")
+                    # Sets the reminders to the state before restart
+                    self.reminder.reminders = rem
 
-                with open("cache/reminders.temp", "rb") as vote_cache:
-                    rem = load(vote_cache)
-
-                # 3.1.5 : disabled
-                # os.remove("cache/reminders.temp")
-
-                # Sets the reminders to the state before restart
-                self.reminder.reminders = rem
+        else:
+            self.reminder = RedisReminderHandler(self.client, self.handler, self.loop)
 
         self.loop.create_task(self.reminder.start_monitoring())
 
@@ -228,11 +388,10 @@ class Reminder:
             else:
                 ttr = int(args[0])
 
-            resp = self.reminder.set_reminder(message.author, message.author, content, ttr)
+            resp = self.reminder.set_reminder(message.author, message.author, content, ttr, reminder_type=REMINDER_PERSONAL)
 
             if resp == -1:
-                await client.send_message(message.channel,
-                                          "You have exceeded the maximum amount of reminders (**1** at once).")
+                await client.send_message(message.channel, EXCEEDED_REMINDER_LIMIT)
 
             elif resp is False:
                 await client.send_message(message.channel, "Not a valid time range (5 seconds to 2 days")
@@ -256,11 +415,10 @@ class Reminder:
             else:
                 ttr = int(args[0])
 
-            resp = self.reminder.set_reminder(message.channel, message.author, content, ttr)
+            resp = self.reminder.set_reminder(message.channel, message.author, content, ttr, reminder_type=REMINDER_CHANNEL)
 
             if resp == -1:
-                await client.send_message(message.channel,
-                                          "You have exceeded the maximum amount of reminders (**1** at once).")
+                await client.send_message(message.channel, EXCEEDED_REMINDER_LIMIT)
 
             elif resp is False:
                 await client.send_message(message.channel, "Not a valid time range (5 seconds to 2 days")
@@ -270,20 +428,25 @@ class Reminder:
 
         # !remind list
         elif startswith(prefix + "remind list", prefix + "reminder list"):
-            reminders = self.reminder.get_reminders(message.author)
+            reminders = self.reminder.get_reminders(message.author.id)
 
             if not reminders:
-                await client.send_message(message.channel, "You have not set any reminders.")
+                await client.send_message(message.channel, "You don't have any reminders.")
                 return
 
             rem = []
-            for reminder in reminders:
+            for reminder in reminders.values():
                 # Gets the remaining time
                 ttl = reminder.get("time_target") - time.time()
 
                 cont = self.nano.get_plugin("commons").get("instance").at_everyone_filter(reminder.get("raw"), message.author, message.server)
 
-                rem.append("➤ {} (in **{}**)".format(cont, resolve_time(ttl)))
+                if (ttl != abs(ttl)) or ttl == 0:
+                    when = "**soon™**"
+                else:
+                    when = "in **{}**".format(resolve_time(ttl))
+
+                rem.append("➤ {} ({})".format(cont, when))
 
             await client.send_message(message.channel, "Your reminders:\n" + "\n".join(rem))
 
@@ -295,7 +458,7 @@ class Reminder:
                 self.reminder.remove_all_reminders(message.author)
 
             else:
-                r_resp = self.reminder.remove_reminder(message.author, r_name)
+                r_resp = self.reminder.remove_reminder(message.author.id, r_name)
 
                 if not r_resp:
                     await client.send_message(message.channel, "No reminder with such content.")
@@ -307,6 +470,9 @@ class Reminder:
             await client.send_message(message.channel, remind_help.replace("_", prefix))
 
     async def on_shutdown(self, **_):
+        if not self.legacy:
+            return
+
         # Saves the state
         if not os.path.isdir("cache"):
             os.mkdir("cache")
@@ -323,8 +489,8 @@ class Reminder:
 
 
 class NanoPlugin:
-    _name = "Reminder Commands"
-    _version = "0.2.3"
+    name = "Reminder Commands"
+    version = "0.2.3"
 
     handler = Reminder
     events = {
